+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog].
   canonical representation of a candidate has been removed ([#403]).
 
 ### Features
+* Line spacing is taken into account when using a fixed window height
+  ([#424], [#432]).
 * The new option `selectrum-max-window-height` can now be used to
   configure the maximal display window height analogue to the built-in
   `max-mini-window-height`. The new option replaces the usage of the
@@ -52,6 +54,9 @@ The format is based on [Keep a Changelog].
   default is t) ([#261]).
 
 ### Enhancements
+* The default value of `selectrum-extend-current-candidate-highlight`
+  has been changed to `auto` which will automatically extend the
+  highlighting if the session uses annotations ([#427], [#430]).
 * The accessibility of the default value in file completions has been
   improved. When the default exists in the prompting directory it gets
   sorted first otherwise the default is included as a virtual
@@ -362,7 +367,11 @@ The format is based on [Keep a Changelog].
 [#413]: https://github.com/raxod502/selectrum/pull/413
 [#414]: https://github.com/raxod502/selectrum/pull/414
 [#420]: https://github.com/raxod502/selectrum/issues/420
+[#424]: https://github.com/raxod502/selectrum/issues/424
 [#421]: https://github.com/raxod502/selectrum/pull/421
+[#427]: https://github.com/raxod502/selectrum/issues/427
+[#430]: https://github.com/raxod502/selectrum/pull/430
+[#432]: https://github.com/raxod502/selectrum/pull/432
 [#438]: https://github.com/raxod502/selectrum/issues/438
 [#439]: https://github.com/raxod502/selectrum/pull/439